requirements.txt
streamlit
paddleocr
paddlepaddle
langchain
langchain-openai
Pillow
numpy
